+++ core/core.pl
@@ -3404,7 +3404,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 % versions, 'rcN' for release candidates (with N being a natural number),
 % and 'stable' for stable versions
 
-'$lgt_version_data'(logtalk(3, 31, 0, b02)).
+'$lgt_version_data'(logtalk(3, 31, 0, b03)).
 
 
 
@@ -8839,8 +8839,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 	;	'$lgt_print_message'(silent(compiling), compiling_entity(object, Obj)),
 		'$lgt_compile_object_relations'(Relations, Obj, Ctx),
 		'$lgt_compile_object_identifier'(Obj, Ctx),
-		'$lgt_comp_ctx_term'(Ctx, Term),
-		'$lgt_save_parameter_variables'(Obj, Term)
+		'$lgt_save_parameter_variables'(Obj, Ctx)
 	).
 
 '$lgt_compile_logtalk_directive'(end_object, Ctx) :-
@@ -8945,8 +8944,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 	;	'$lgt_print_message'(silent(compiling), compiling_entity(category, Ctg)),
 		'$lgt_compile_category_identifier'(Ctg, Ctx),
 		'$lgt_compile_category_relations'(Relations, Ctg, Ctx),
-		'$lgt_comp_ctx_term'(Ctx, Term),
-		'$lgt_save_parameter_variables'(Ctg, Term)
+		'$lgt_save_parameter_variables'(Ctg, Ctx)
 	).
 
 '$lgt_compile_logtalk_directive'(end_category, Ctx) :-
@@ -16687,8 +16685,8 @@ create_logtalk_flag(Flag, Value, Options) :-
 
 
 
-% '$lgt_save_parameter_variables'(@object_identifier, @callable)
-% '$lgt_save_parameter_variables'(@category_identifier, @callable)
+% '$lgt_save_parameter_variables'(@object_identifier, @compilation_context)
+% '$lgt_save_parameter_variables'(@category_identifier, @compilation_context)
 %
 % saves the parameter variable names and positions found
 % in parametric entity identifiers for later processing
@@ -16696,11 +16694,13 @@ create_logtalk_flag(Flag, Value, Options) :-
 % the original term is passed to the predicate to unify the variables
 % in the entity with the variables in the variable names list
 
-'$lgt_save_parameter_variables'(Entity, Term) :-
+'$lgt_save_parameter_variables'(Entity, Ctx) :-
 	compound(Entity),
+	'$lgt_comp_ctx_term'(Ctx, Term),
 	'$lgt_pp_term_variable_names_file_lines_'(Term, VariableNames, _, _),
-	Entity =.. [_| Arguments],
-	'$lgt_parameter_variable_pairs'(VariableNames, Arguments, ParameterVariablePairs),
+	Entity =.. [_| Parameters],
+	'$lgt_check_for_nonvar_parameters'(Parameters, Ctx),
+	'$lgt_parameter_variable_pairs'(VariableNames, Parameters, ParameterVariablePairs),
 	ParameterVariablePairs \== [],
 	!,
 	assertz('$lgt_pp_parameter_variables_'(ParameterVariablePairs)).
@@ -16708,16 +16708,28 @@ create_logtalk_flag(Flag, Value, Options) :-
 '$lgt_save_parameter_variables'(_, _).
 
 
+'$lgt_check_for_nonvar_parameters'(Parameters, Ctx) :-
+	(	'$lgt_comp_ctx_mode'(Ctx, runtime) ->
+		true
+	;	'$lgt_member'(Parameter, Parameters),
+		nonvar(Parameter) ->
+		'$lgt_increment_compiling_warnings_counter',
+		'$lgt_source_file_context'(File, Lines, Type, Entity),
+		'$lgt_print_message'(warning(general), nonvar_parameter(File, Lines, Type, Entity, Parameter))
+	;	true
+	).
+
+
 '$lgt_parameter_variable_pairs'([], _, []).
 
-'$lgt_parameter_variable_pairs'([VariableName=Variable| VariableNames], Arguments, [VariableName-Position| ParameterVariablePairs]) :-
+'$lgt_parameter_variable_pairs'([VariableName=Variable| VariableNames], Parameters, [VariableName-Position| ParameterVariablePairs]) :-
 	'$lgt_parameter_variable_name'(VariableName),
-	'$lgt_parameter_variable_position'(Arguments, Variable, 1, Position),
+	'$lgt_parameter_variable_position'(Parameters, Variable, 1, Position),
 	!,
-	'$lgt_parameter_variable_pairs'(VariableNames, Arguments, ParameterVariablePairs).
+	'$lgt_parameter_variable_pairs'(VariableNames, Parameters, ParameterVariablePairs).
 
-'$lgt_parameter_variable_pairs'([_| VariableNames], Arguments, ParameterVariablePairs) :-
-	'$lgt_parameter_variable_pairs'(VariableNames, Arguments, ParameterVariablePairs).
+'$lgt_parameter_variable_pairs'([_| VariableNames], Parameters, ParameterVariablePairs) :-
+	'$lgt_parameter_variable_pairs'(VariableNames, Parameters, ParameterVariablePairs).
 
 
 % '$lgt_parameter_variable_name'(+atom)